--- src/main/java/arom/springtoy/HomeController.java
@@ -1,5 +1,0 @@
-package arom.springtoy;
-
-public class HomeController {
-
-}
