--- src/main/java/arom/springtoy/InitData.java
@@ -1,71 +1,0 @@
-package arom.springtoy;
-
-import arom.springtoy.content.domain.Content;
-import arom.springtoy.content.repository.ContentRepository;
-import arom.springtoy.todolist.domain.Todolist;
-import arom.springtoy.todolist.dto.DateDto;
-import arom.springtoy.todolist.repository.TodolistRepository;
-import arom.springtoy.user.domain.User;
-import arom.springtoy.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.context.event.EventListener;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
-@Component
-@RequiredArgsConstructor
-public class InitData {
-
-    private final UserRepository userRepository;
-    private final TodolistRepository todolistRepository;
-    private final ContentRepository contentRepository;
-
-    @EventListener(ApplicationReadyEvent.class)
-    @Transactional
-    public void init(){
-        DateDto startDate = new DateDto(
-            2024,3,6,14,23
-        );
-        DateDto endDate = new DateDto(
-            2024,4,20,15,30
-        );
-
-        User user1 = new User(
-            "inittest1@naver.com",
-            "123",
-            "initTestNickname1"
-        );
-        User user2 = new User(
-            "inittest2@naver.com",
-            "1234",
-            "initTestNickname2"
-        );
-
-        userRepository.save(user1);
-        userRepository.save(user2);
-
-        Todolist todolist1 = new Todolist(
-            user1,
-            "testWriter1",
-            "testTodolistName1",
-            startDate,
-            endDate
-        );
-
-        todolistRepository.save(todolist1);
-
-        Content content1 = new Content(
-            todolist1,
-            "testContent1",
-            "write down what to do"
-        );
-        Content content2 = new Content(
-            todolist1,
-            "testContent2",
-            "write down what to do"
-        );
-        contentRepository.save(content1);
-        contentRepository.save(content2);
-    }
-}
