+++ src/main/java/arom/springtoy/domain/dto/JoinDto.java
@@ -1,0 +1,29 @@
+package arom.springtoy.domain.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class JoinDto {
+    @Email
+    private String email;
+
+    @NotNull
+    @NotBlank
+    private String password;
+
+    @NotNull @NotBlank
+    @Size(min = 1, max = 30)
+    private String nickname;
+
+}
