+++ src/main/java/arom/springtoy/domain/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package arom.springtoy.domain.repository;
+
+import arom.springtoy.domain.domain.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String Email);
+    Optional<User> findByNickname(String nickname);
+}
