--- src/main/java/arom/springtoy/todolist/service/ListGenerator.java
@@ -1,5 +1,0 @@
-package arom.springtoy.todolist.service;
-
-public class ListGenerator {
-
-}
