--- src/main/java/arom/springtoy/user/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package arom.springtoy.user.repository;
-
-import arom.springtoy.user.domain.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(String Email);
-    Optional<User> findByNickname(String nickname);
-}
